+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 1.2.29
 - Made cloak combine alpha when taking off / landing
+- Fixed particles not showing